Use OpenShift Ansible modules instead of templates
Currently we are creating postgres using a template. We copy the
template onto the host machine and then perform an oc create. This
does not take advantage of the Ansible modules which are avaiable.

This change replaces the template yaml file with a set of Ansible
tasks. It also replaces the jar file for Keycloak Monitoring
Prometheus to use the latest upstream version instead of a forked
version of the source code. This new version allows an environment
variable to be specified for the Keycloak metrics directory instead
of having to modify the Keycloak configuration files.