Make keycloak protocol configurable
Currently our keycloak protocol is https only. If for any reason
developers need to use http this is not ideal and requires a lot
of changes to move between the two.

This change makes the keycloak protocol configurable through the
keycloak_protocol variable. When set to http then a http route will
be create, when https then a https route will be created.

This variable is not a APB variable and is not available to an
everyday user. Developers will need to change the variable and
rebuild the APB to change the protocol.