Remove clientID from configuration
Currently we include a clientID field in the secret for Keycloak
that gets converted by the Mobile CLI. This results in there being
two keys with the same value. 'resource' which comes from Keycloak
and gives the identifier of the client and 'clientId' which is
appeneded by the APB and gives the same identifier without
modification.

This change removes the duplicate key.